Property implementation
redesigned the multiple-class approach to a single class
because xmlSerialization was not able to deserialize it
according to S3000L schemas.
Since the S3000L does not store an xsi:type it is hard
for the xmlSerialization to detect the correct Property-
Implementation.
Implementing IXmlSerializable would have been a pain in ass,
because I would have to implement it on the containing class
and not on the Property-Base Class :/